Refactored the Gesture enum from Paj7620.cs into a own file and added intellisense description to the Gestures.
